moved telegram channel methods to telegram controller
